Bug fix
- Newline deskripsi produk sesuai dengan yang di tulis Admin
- Notif angka jumlah barang yg ada di keranjang
- Daftar pembelian oleh user (tab semua) urutkan dari yang belum bayar -> lacak pengiriman -> beli lagi
- Error handling checkout sequence
- Upload bukti pembayaran -> field amount jadi bigInteger
- (Admin) Gambar produk dimunculkan saat mau edit

+++ database/migrations/2020_12_06_200239_create_payments_table.php
@@ -22,7 +22,7 @@ class CreatePaymentsTable extends Migration
             $table->string('transfer_from_bank')->nullable();
             $table->string('transfer_from_account')->nullable();
             $table->date('transfer_date')->nullable();
-            $table->integer('amount')->nullable();
+            $table->bigInteger('amount')->nullable();
             $table->integer('status')->default(0);
             $table->timestamps();
         });
